sqlite with traceを別ファイルに逃がす
